+++ tests/unit/cli/test_main.py
@@ -39,7 +39,7 @@ def test_state_pickle_errors_are_correctly_raised(tmp_dir, caplog, mocker):
 )
 def test_remote_missing_deps_are_correctly_reported(tmp_dir, caplog, mocker, pkg, msg):
     error = RemoteMissingDepsError(FileSystem(), "proto", "proto://", ["deps"])
-    mocker.patch("dvc.utils.pkg.PKG", pkg)
+    mocker.patch("dvc.PKG", pkg)
     mocker.patch(
         "dvc.cli.parse_args",
         return_value=Namespace(
